+++ lab_03_v2/lab_03_final/lab_03/solution/solution.cpp
@@ -1,22 +1,22 @@
 #include "errors.hpp"
 #include "solution.hpp"
 
-bool Solution::registration_creator(size_t index, const std::shared_ptr<LoaderCreator> &lc)
+bool Solution::registration_creator(int index, const std::shared_ptr<LoaderCreator> &lc)
 {
     return creator_map.insert(callback_creator::value_type(index, lc)).second;
 }
 
-bool Solution::registration_scene_creator(size_t index, const std::shared_ptr<SceneLoaderCreator> &lc)
+bool Solution::registration_scene_creator(int index, const std::shared_ptr<SceneLoaderCreator> &lc)
 {
     return scene_map.insert(callback_scene_creator::value_type(index, lc)).second;
 }
 
-bool Solution::registration_factory(size_t index, const std::shared_ptr<AbstractFactory> &lc)
+bool Solution::registration_factory(int index, const std::shared_ptr<AbstractFactory> &lc)
 {
     return factory_map.insert(callback_factory::value_type(index, lc)).second;
 }
 
-std::shared_ptr<LoaderCreator> Solution::get_creator(size_t index)
+std::shared_ptr<LoaderCreator> Solution::get_creator(int index)
 {
     callback_creator::const_iterator it = creator_map.find(index);
     if (it == creator_map.end())
@@ -28,7 +28,7 @@ std::shared_ptr<LoaderCreator> Solution::get_creator(size_t index)
     return it->second;
 }
 
-std::shared_ptr<SceneLoaderCreator> Solution::get_scene_creator(size_t index)
+std::shared_ptr<SceneLoaderCreator> Solution::get_scene_creator(int index)
 {
     callback_scene_creator::const_iterator it = scene_map.find(index);
     if (it == scene_map.end())
@@ -40,7 +40,7 @@ std::shared_ptr<SceneLoaderCreator> Solution::get_scene_creator(size_t index)
     return it->second;
 }
 
-std::shared_ptr<AbstractFactory> Solution::get_factory(size_t index)
+std::shared_ptr<AbstractFactory> Solution::get_factory(int index)
 {
     callback_factory::const_iterator it = factory_map.find(index);
     if (it == factory_map.end())
